apps/streamlit/requirements.txt
# first install dependent pkgs from 
# $PROMPTFIX_HOME/requirements.txt
streamlit
streamlit-aggrid
pandas
torch 
einops 
omegaconf 
pillow 
k-diffusion
